Back to CommonJS
The release using ES2015 modules broke both npm and yarn! Whaaa?

This structures the project in such a way it should be friendlier to CommonJS
